tests/wpt/tests/css/css-backgrounds/parsing/border-image-outset-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-outset with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-outset">
<meta name="assert" content="border-image-outset supports only the grammar '[ <length> | <number> ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("border-image-outset", "auto");

test_invalid_value("border-image-outset", "-1");
test_invalid_value("border-image-outset", "-2px");

test_invalid_value("border-image-outset", "1%");

test_invalid_value("border-image-outset", "1 2 3 4 5");
</script>
</body>
</html>
